refactor: 오타 수정 및 불필요한 함수 호출 제거
- isPageRefreshing 오타 수정
- reloadData 제거
- indexPathsForVisibleItems를 이용해 현재 보이는 cell만 reloadItem 진행
- 반환하는 item 수를 items.count로 설정

+++ OpenMarket/OpenMarket/Model/JSON/Item.swift
@@ -11,9 +11,9 @@ struct Item: Codable, Hashable {
     let name: String
     let thumbnail: String
     let currency: String
-    let price: Int
-    let bargainPrice: Int
-    let discountedPrice: Int
+    let price: Double
+    let bargainPrice: Double
+    let discountedPrice: Double
     let stock: Int
     let createdAt: String
     let issuedAt: String
